+++ packages/satcheljs/lib/mutator.ts
@@ -1,7 +1,7 @@
 import ActionCreator from './interfaces/ActionCreator';
 import ActionMessage from './interfaces/ActionMessage';
 import Subscriber from './interfaces/Subscriber';
-import { subscribe } from './subscriptions';
+import { subscribe } from './dispatcher';
 
 export default function mutator<T extends ActionMessage>(actionDispatcher: ActionCreator<T>) {
     return function subscribeMutator(callback: Subscriber<T>) {
